+++ bugsnag-android-core/src/sharedTest/java/com/bugsnag/android/BugsnagTestUtils.java
@@ -62,7 +62,7 @@ final class BugsnagTestUtils {
 
     static Session generateSession() {
         return new Session("test", new Date(), new User(), false,
-                new Notifier(), NoopLogger.INSTANCE);
+                new Notifier(), NoopLogger.INSTANCE, "BUGSNAG_API_KEY");
     }
 
     static Event generateEvent() {